split to files and refactor
