+++ README.md
@@ -1,3 +1,5 @@
 ## assignment_mastermind
 
 [An object-oriented Ruby assignment from the Viking Code School](http://www.vikingcodeschool.com)
+
+> -- Pramod & Sampson
